Changes to the MainMenu: - Added the functionality to use the mouse to select and trigger buttons - Moved Button Class to a seperate file - Added comments to the code
