+++ GameBox/GameBox/GameBox.vcxproj
@@ -174,6 +174,7 @@
     <ClCompile Include="ECS\Systems\SpriteRenderSystem.cpp" />
     <ClCompile Include="ECS\Systems\TextSystem.cpp" />
     <ClCompile Include="Game.cpp" />
+    <ClCompile Include="state\Common UI\Button.cpp" />
     <ClCompile Include="state\GameState.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="state\MainMenuState.cpp">
@@ -199,6 +200,7 @@
     <ClInclude Include="events\ClickActionEvent.h" />
     <ClInclude Include="events\ScoreEvent.h" />
     <ClInclude Include="Game.h" />
+    <ClInclude Include="state\Common UI\Button.h" />
     <ClInclude Include="state\GameState.h" />
     <ClInclude Include="state\MainMenuState.h" />
     <ClInclude Include="state\State.h" />
