+++ GameBox/GameBox/state/MainMenuState.cpp
@@ -48,17 +48,25 @@ void MainMenuState::processInput(float dt) {
 }
 
 void MainMenuState::handleWindowEvent(const sf::Event& windowEvent) {
+	// switch statement to handle different types of window events
 	switch (windowEvent.type) {
+
+		// case for when a key is pressed
 	case sf::Event::EventType::KeyPressed:
+		// if the Enter key is pressed, trigger the currently selected button
 		if (windowEvent.key.code == sf::Keyboard::Enter) {
 			m_buttons[m_selectedButton].Trigger();
 		}
+
+		// if the W or Up arrow key is pressed, move the selection up to the previous button if it exists
 		if (windowEvent.key.code == sf::Keyboard::W || windowEvent.key.code == sf::Keyboard::Up) {
 			if (m_selectedButton > 0) {
 				m_buttons[m_selectedButton--].SetIsSelected(false);
 				m_buttons[m_selectedButton].SetIsSelected(true);
 			}
 		}
+
+		// if the S or Down arrow key is pressed, move the selection down to the next button if it exists
 		if (windowEvent.key.code == sf::Keyboard::S || windowEvent.key.code == sf::Keyboard::Down) {
 			if (m_selectedButton < m_buttons.size() - 1) {
 				m_buttons[m_selectedButton++].SetIsSelected(false);
@@ -66,41 +74,50 @@ void MainMenuState::handleWindowEvent(const sf::Event& windowEvent) {
 			}
 		}
 		break;
+
+		// case for when a mouse button is pressed
+	case sf::Event::MouseButtonPressed:
+	{
+		// get the ID of the button currently being hovered over
+		size_t buttonID = getHoveredButtonID();
+
+		// if a button is being hovered over, trigger it
+		if (buttonID != -1) {
+			m_buttons[buttonID].Trigger();
+		}
+		break;
+	}
+		// case for when the mouse is moved
+	case sf::Event::EventType::MouseMoved:
+	{
+		// get the ID of the button currently being hovered over
+		size_t buttonID = getHoveredButtonID();
+
+		// if a button is being hovered over, set it as the selected button
+		if (buttonID != -1) {
+			m_buttons[m_selectedButton].SetIsSelected(false);
+			m_buttons[buttonID].SetIsSelected(true);
+			m_selectedButton = buttonID;
+		}
+		break;
+	}
+	// default case for handling any other event types not covered in the switch statement
 	default:
 		break;
 	}
+
 }
 
-Button::Button(sf::String text, const sf::Font& font, const std::function<void()> func) : m_func(func) {
-	m_text.setFont(font);
-	m_text.setString(text);
-}
+size_t MainMenuState::getHoveredButtonID() {
 
-Button::~Button() {
-}
-
-void Button::Trigger() {
-	m_func();
-}
-
-void Button::SetIsSelected(bool b) {
-	m_isSelected = b;
-	if (m_isSelected) {
-		m_text.setStyle(sf::Text::Bold | sf::Text::Italic);
-	} else {
-		m_text.setStyle(0);
+	size_t id = 0;
+	for (auto& i : m_buttons) {
+		sf::Vector2i mousePos = sf::Mouse::getPosition(*m_game->getWindow());
+		if (i.GetRectangle().contains(mousePos.x, mousePos.y)) {
+			return id;
+		}
+		id++;
 	}
-}
 
-void Button::setPosition(float x, float y) {
-	m_text.setPosition(x, y);
-}
-
-void Button::draw(sf::RenderTarget& target, sf::RenderStates states) const {
-	if (m_isSelected) {
-
-	} else {
-
-	}
-	target.draw(m_text, states);
+	return -1;
 }
