+++ GameBox/GameBox/state/MainMenuState.h
@@ -4,23 +4,7 @@
 #include <iostream>
 #include "State.h"
 #include <functional>
-
-class Button : public sf::Drawable {
-public:
-	Button(sf::String text, const sf::Font& font, const std::function<void()> func);
-	~Button();
-	void Trigger();
-	void SetIsSelected(bool b);
-	void setPosition(float x, float y);
-
-private:
-	// Inherited via Drawable
-	virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const override;
-
-	std::function<void()> m_func;
-	bool m_isSelected = false;
-	sf::Text m_text;
-};
+#include "Common UI/Button.h"
 
 class MainMenuState final : public State {
 public:
@@ -31,6 +15,11 @@ public:
 	void processInput(float dt) override;
 	virtual void handleWindowEvent(const sf::Event& windowEvent) override;
 private:
+	/*
+		Returns the ID of the button currently being hovered over by the mouse, or -1 if no button is being hovered over
+	*/
+	size_t getHoveredButtonID();
+
 	size_t m_selectedButton = 0;
 	std::vector<Button> m_buttons;
